👔 #18 List 컴포넌트 추가 및 Grid 컴포넌트 분리
